// packages/datadog-api-client-v1/models/SyntheticsSSLCertificateIssuer.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Object describing the issuer of a SSL certificate.
*/
export declare class SyntheticsSSLCertificateIssuer {
    /**
    * Country Name that issued the certificate.
    */
    'C'?: string;
    /**
    * Common Name that issued certificate.
    */
    'CN'?: string;
    /**
    * Locality that issued the certificate.
    */
    'L'?: string;
    /**
    * Organization that issued the certificate.
    */
    'O'?: string;
    /**
    * Organizational Unit that issued the certificate.
    */
    'OU'?: string;
    /**
    * State Or Province Name that issued the certificate.
    */
    'ST'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsSSLCertificateIssuer;
    static serialize(data: SyntheticsSSLCertificateIssuer): {
        [key: string]: any;
    };
    constructor();
}