// packages/datadog-api-client-v1/models/SyntheticsSSLCertificateSubject.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Object describing the SSL certificate used for the test.
*/
export declare class SyntheticsSSLCertificateSubject {
    /**
    * Country Name associated with the certificate.
    */
    'C'?: string;
    /**
    * Common Name that associated with the certificate.
    */
    'CN'?: string;
    /**
    * Locality associated with the certificate.
    */
    'L'?: string;
    /**
    * Organization associated with the certificate.
    */
    'O'?: string;
    /**
    * Organizational Unit associated with the certificate.
    */
    'OU'?: string;
    /**
    * State Or Province Name associated with the certificate.
    */
    'ST'?: string;
    /**
    * Subject Alternative Name associated with the certificate.
    */
    'altName'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SyntheticsSSLCertificateSubject;
    static serialize(data: SyntheticsSSLCertificateSubject): {
        [key: string]: any;
    };
    constructor();
}